Split runner file into multiple files
